DRAFT PLAYBOOK USED TO CREATE THE BACKUP, DNS, MONITOR AND WEB SERVICES
The create-services.yml playbook here has been used in conjunction with
the Ansible roles in /roles and the Varilink Computing Ltd Ansible
inventory to create our initial backup, dns, monitor and web services.
A very brief README.md update is also included as well as update to the
latest commit point of the roles submodule.